build(flake): ensure that nix flake have deterministic outputs
TODO:
Install dependencies declared in `pyproject.toml` through nix
instead of pip so as to pin the python dev dependencies. The old way
to do this is pinning them in `requirements-dev.txt` generated by
pip-compile and reproducing with pip-sync in a virtual environment.

After doing this, build this lib as a nix python package.

Bibliography:
- https://github.com/ryan4yin/nixos-and-flakes-book/blob/e30b9a44f27dbacb7415f4449b0353e2a94f3b01/docs/zh/development/dev-environments.md?plain=1#L54-L66
- https://github.com/nix-community/pyproject.nix
  - https://github.com/nix-community/pyproject.nix/blob/5edc794d76ec64eb699f75d65fd19299ee42c99e/templates/pyproject/flake.nix#L53-L63